fix(orchestrator): open a disconnected remote from the dock non-blockingly
Selecting a not-yet-connected remote (SSH/Kubernetes) workspace in the dock —
by arrow-nav or click — must open it with a clear "Connecting…" message in
the workspace area while the connect runs in the background, and must never be
prevented or freeze. An earlier attempt skipped the arrow-nav switch entirely
for disconnected remotes; that stopped the freeze but also stopped you from
opening them, which isn't what's wanted.

Restore the arrow-nav live-switch for these rows: the host already commits
into a placeholder "Connecting…" page and connects asynchronously
(`ensure_dormant_shell` + `bring_dormant_remote_online`, the #2570 path), so
the switch is non-blocking, and the 30 ms debounce already means only pausing
on a row (not scrolling past it) triggers it. Keep the `dock_activate` (Enter)
change so Enter still commits the switch if it beats the debounce.

Adds `orchestrator_dock_ssh_navigation_nonblocking`: arrowing onto a dormant
SSH row whose host hangs commits into the "Connecting…" placeholder page (a
visible message, never a frozen buffer) without blocking. Restores the
dormant-SSH badge test to its arrow-nav-commits form.
